+++ src/extension.ts
@@ -2,9 +2,10 @@ import * as vscode from 'vscode';
 
 import { ICommand } from './lib/types';
 import { info, outputChannel } from './lib/outputPane';
-import { LinkGitConfig } from './commands/link-git-config';
+import { LinkConfigGit } from './commands/link-config-git';
 import { LinkInit } from './commands/link-init';
 import { LinkTest } from './commands/link-test';
+import { LinkValidateDecisions } from './commands/link-validate-decisions';
 import { LogReplay } from './commands/log-replay';
 
 // extension entry point
@@ -15,8 +16,9 @@ export function activate(context: vscode.ExtensionContext): void {
   // commands come from package.json::contribute.commands
   const styraCommands: { [key: string]: ICommand } = {
     'styra.link.init': new LinkInit(),
-    'styra.link.git-config': new LinkGitConfig(),
+    'styra.link.config-git': new LinkConfigGit(),
     'styra.link.test': new LinkTest(),
+    'styra.link.validate-decisions': new LinkValidateDecisions(),
     'styra.log.replay': new LogReplay()
   };